two-second long press clear all timers; button tap stops running timer
Pressing and holding the clear button for two seconds will clear
all timers and stop all watches.  Times are visually reset
to zero and zero percent, as well.  (This required a new reset
function to be added to the stopwatch that sets all variables to
their initial values.)

A timer that is currently running will stop when the button is hit.
Original performance would only start a stopped timer and do nothing
for an already running timer.