new version; upped font in main labels
I added a bunch of new ignore files.

The new version reflects changes to the font size and replacing
all dudes with bros in the labels.

The font changes are relative to the size of the buttons and seem
to work fairly well for phone screen size(s) and a bunch of
reasonable different terminal run sizes.