added about screen and rejuggled buttons accordingly
The 'save' button has been removed (for now) to make space
for an 'about' button.  The 'about' button takes you to a new
screen, which necessitated the introduction of a screen manager
and renaming some variables to match.  The about screen
stops/resets all timers so I disabled the button when any
buttons have meaningful numbers.  (That is, you can only click
about when you first enter the program or after you have cleared
the last results.)