simple save feature on button long press
When the timers are all zero, then the left-middle button will
take you to the about screen.  When there is any time on any
timer, then the about button turns in to a save button.  Long
pressing the save (2secs like with clear) will write all
times to file:  youdonebro_YYYY-MM-DD_HHMMSS.csv