fixed dec button typo; pushed inc/dec out a bit for new count lbl
I moved the counts to a new label inbetween the inc and dec
buttons.  I noticed a typo on one dec button that showed a zero
instead of a -.

Change the color of the back-to-app button in the about screen to
match the color scheme of the rest of the buttons.